requirements.txt
confusable-homoglyphs==2.0.2
Django==2.0
django-registration==2.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
